basic traveldiary screen functionality and react navigation installation
