+++ frontend/package.json
@@ -10,12 +10,19 @@
     "test": "echo \"No tests available\" && exit 0"
   },
   "dependencies": {
+    "@react-navigation/native": "^7.0.14",
+    "@react-navigation/stack": "^7.1.1",
     "expo": "~52.0.31",
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
     "react-native": "0.76.7",
     "react-native-elements": "^3.4.3",
-    "react-native-maps": "^1.20.1"
+    "react-native-gesture-handler": "~2.20.2",
+    "react-native-maps": "^1.20.1",
+    "react-native-reanimated": "~3.16.1",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.4.0",
+    "react-native-vector-icons": "^10.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
